requirements.txt
Pillow>=7.0.0
pillow-heif>=0.13
